# perf: lazy load babel transform