--- a/lib/jiti.mjs
+++ b/lib/jiti.mjs
@@ -1,6 +1,5 @@
 import { createRequire } from "node:module";
 import _createJiti from "../dist/jiti.cjs";
-import transform from "../dist/babel.cjs";
 
 function onError(err) {
   throw err; /* ↓ Check stack trace ↓ */
@@ -8,9 +7,18 @@ function onError(err) {
 
 const nativeImport = (id) => import(id);
 
+let _transform;
+function lazyTransform(...args) {
+  if (!_transform) {
+    const { require } = createRequire(import.meta.url);
+    _transform = require("../dist/babel.cjs");
+  }
+  return _transform(...args);
+}
+
 export function createJiti(id, opts = {}) {
   if (!opts.transform) {
-    opts = { ...opts, transform };
+    opts = { ...opts, transform: lazyTransform };
   }
   return _createJiti(id, opts, {
     onError,